fix(dashboard): fold monolith into host + default tiles
Retire the core:dashboard monolith (fold + delete, issue #133). The
dashboard screen now mounts core:dashboard-host with a bundled default
tile set; the greeting becomes host chrome (a header above the grid),
not a tile.

- Delete src/apps/dashboard/ + its core:dashboard builtins registration.
- Repoint the empty-doc baseline + docs from core:dashboard to
  core:dashboard-host.
- Decompose the monolith into default tiles:
  - core:dashboard-widget-at-a-glance (NEW) — site-wide counts.
  - core:dashboard-widget-activity (NEW) — recently published + pending
    comments (site-wide).
  - core:dashboard-widget-recent-posts — Recent Drafts, now author-scoped
    (author: userId + enabled: !!userId, fail-closed; carries the #217
    draft-leak fix from the monolith).
  - core:dashboard-widget-quick-draft — promoted to a default tile;
    invalidates the author-scoped recent-drafts query.
- query.mjs now exports recentDraftsQuery(userId) so the Recent Drafts
  read and the Quick Draft invalidation share one author-scoped shape.
- Greeting header added to the host (time-of-day + acting user name).
- wp-admin-default dashboard-home mounts all four tiles.
- Update dashboard-host app.json#documentation + app.md; code-map.
