+++ src/apps/dashboard-widget-activity/app.json
@@ -1,0 +1,81 @@
+{
+	"$schema": "../../../docs/schemas/admin-app.json",
+	"id": "core:dashboard-widget-activity",
+	"version": 1,
+	"title": "Activity",
+	"description": "Recently published posts plus comments awaiting moderation.",
+	"designSystem": "@wordpress/ui",
+	"role": "region",
+	"capabilities": [ "edit_posts" ],
+	"config-schema": {
+		"type": "object",
+		"additionalProperties": false
+	},
+	"script": "wp-admin-shell",
+	"slotHints": {
+		"defaultSize": { "w": 2, "h": 1 },
+		"minSize": { "w": 1, "h": 1 },
+		"position": "auto"
+	},
+	"documentation": {
+		"purpose": "Mirrors wp-admin's \"Activity\" box: the five most recently published posts plus comments awaiting moderation. Both lists are site-wide (only Recent Drafts is author-scoped per issue #133).",
+		"data": {
+			"reads": [
+				{
+					"source": "postType/post",
+					"via": "core-data",
+					"context": "edit",
+					"purpose": "Five most recently published posts (status: publish, orderby: date desc).",
+					"query": { "status": "publish", "per_page": 5, "context": "edit", "orderby": "date", "order": "desc" }
+				},
+				{
+					"source": "root/comment",
+					"via": "core-data",
+					"context": "edit",
+					"purpose": "Five comments awaiting moderation (status: hold).",
+					"query": { "status": "hold", "per_page": 5, "context": "edit" }
+				}
+			]
+		},
+		"url": {
+			"navigates": [ "#/posts/{id}/edit", "#/comments" ]
+		},
+		"states": [
+			{
+				"id": "loading",
+				"when": "posts.isResolving && !posts.records (per section)",
+				"renders": "Spinner in the affected section."
+			},
+			{
+				"id": "empty",
+				"when": "a section's records.length === 0",
+				"renders": "`Nothing published yet.` / `Inbox zero. Nothing pending.` copy."
+			},
+			{
+				"id": "ready",
+				"when": "records present",
+				"renders": "Recently-published list + pending-comments list with a `Moderate all` link."
+			}
+		],
+		"interactions": [
+			{
+				"trigger": "Click a published-post title",
+				"effect": "`navigate('#/posts/{id}/edit')`."
+			},
+			{
+				"trigger": "Click `Moderate all`",
+				"effect": "`navigate('#/comments')`."
+			}
+		],
+		"design-system-leakage": [
+			{
+				"import": "@wordpress/ui#Button, Stack, Text",
+				"purpose": "List + section layout."
+			},
+			{
+				"import": "@wordpress/components#Spinner",
+				"purpose": "Loading state — no WPDS 0.12 port."
+			}
+		]
+	}
+}
