+++ src/apps/dashboard-widget-at-a-glance/app.json
@@ -1,0 +1,77 @@
+{
+	"$schema": "../../../docs/schemas/admin-app.json",
+	"id": "core:dashboard-widget-at-a-glance",
+	"version": 1,
+	"title": "At a Glance",
+	"description": "Site-wide counts — published posts, published pages, pending comments, registered users.",
+	"designSystem": "@wordpress/ui",
+	"role": "region",
+	"capabilities": [ "read" ],
+	"config-schema": {
+		"type": "object",
+		"additionalProperties": false
+	},
+	"script": "wp-admin-shell",
+	"slotHints": {
+		"defaultSize": { "w": 2, "h": 1 },
+		"minSize": { "w": 1, "h": 1 },
+		"position": "auto"
+	},
+	"documentation": {
+		"purpose": "Mirrors wp-admin's \"At a Glance\" box — aggregate, site-wide counts. Counts are intentionally NOT author-scoped (only Recent Drafts is, per issue #133's design note).",
+		"data": {
+			"reads": [
+				{
+					"source": "postType/post",
+					"via": "core-data",
+					"context": "edit",
+					"purpose": "Published-post count via totalItems (per_page: 1, _fields: id).",
+					"query": { "status": "publish", "per_page": 1, "_fields": "id", "context": "edit" }
+				},
+				{
+					"source": "postType/page",
+					"via": "core-data",
+					"context": "edit",
+					"purpose": "Published-page count via totalItems.",
+					"query": { "status": "publish", "per_page": 1, "_fields": "id", "context": "edit" }
+				},
+				{
+					"source": "root/comment",
+					"via": "core-data",
+					"context": "edit",
+					"purpose": "Pending-comment count via totalItems (status: hold).",
+					"query": { "status": "hold", "per_page": 1, "_fields": "id", "context": "edit" }
+				},
+				{
+					"source": "root/user",
+					"via": "core-data",
+					"purpose": "Registered-user count via totalItems.",
+					"query": { "per_page": 1, "_fields": "id" }
+				}
+			]
+		},
+		"states": [
+			{
+				"id": "loading",
+				"when": "any query isResolving",
+				"renders": "Spinner in the affected stat slot."
+			},
+			{
+				"id": "ready",
+				"when": "queries resolved",
+				"renders": "Four labelled counts (posts / pages / comments / users)."
+			}
+		],
+		"interactions": [],
+		"design-system-leakage": [
+			{
+				"import": "@wordpress/ui#Stack, Text",
+				"purpose": "Stat layout."
+			},
+			{
+				"import": "@wordpress/components#Spinner",
+				"purpose": "Loading state — no WPDS 0.12 port."
+			}
+		]
+	}
+}
