--- src/apps/dashboard/app.json
@@ -1,165 +1,0 @@
-{
-	"$schema": "../../../docs/schemas/admin-app.json",
-	"id": "core:dashboard",
-	"version": 1,
-	"title": "Dashboard",
-	"description": "Site overview cards: recent posts, drafts, comments, quick links.",
-	"designSystem": "@wordpress/ui",
-	"role": "main",
-	"config-schema": {
-		"type": "object",
-		"additionalProperties": false
-	},
-	"script": "wp-admin-shell",
-	"window": {
-		"defaultSize": {
-			"w": 1024,
-			"h": 720
-		},
-		"minSize": {
-			"w": 540,
-			"h": 400
-		},
-		"multiInstance": false,
-		"icon": "wordpress"
-	},
-	"documentation": {
-		"purpose": "Landing screen for authenticated users. Greets the acting user by time-of-day + first name, surfaces three quick-action buttons (write a post, add a page, upload media), renders a stat row (published posts / pages / pending comments / users) and two list cards (recent drafts, comments awaiting moderation). Each count uses a per-status `?per_page=1` REST request reading `X-WP-Total` via `useEntityRecords.totalItems` — WordPress core has no `wp_count_posts` REST equivalent.",
-		"rebuilds": "dashboard-home",
-		"data": {
-			"reads": [
-				{
-					"source": "root/user/{userId}",
-					"via": "core-data",
-					"context": "edit",
-					"purpose": "Greeting personalization (name / first_name)."
-				},
-				{
-					"source": "postType/post",
-					"via": "core-data",
-					"context": "edit",
-					"purpose": "Published-post count (per_page: 1, status: publish, _fields: id) + draft list (per_page: 5, status: draft, orderby: modified)."
-				},
-				{
-					"source": "postType/page",
-					"via": "core-data",
-					"context": "edit",
-					"purpose": "Published-page count."
-				},
-				{
-					"source": "root/comment",
-					"via": "core-data",
-					"context": "edit",
-					"purpose": "Pending-comments count + first 5 pending for the moderation card (status: hold).",
-					"query": {
-						"status": "hold",
-						"per_page": 5,
-						"context": "edit"
-					}
-				},
-				{
-					"source": "root/user",
-					"via": "core-data",
-					"purpose": "Users count.",
-					"query": {
-						"per_page": 1,
-						"_fields": "id"
-					}
-				},
-				{
-					"source": "window.wpAdminShell.userId",
-					"via": "window-global",
-					"purpose": "Acting user id for personalization."
-				}
-			]
-		},
-		"url": {
-			"navigates": [
-				"#/posts/new",
-				"#/pages/new",
-				"#/media",
-				"#/comments",
-				"#/posts/{id}/edit"
-			]
-		},
-		"states": [
-			{
-				"id": "loading-stats",
-				"when": "any of the count queries is resolving",
-				"renders": "Per-stat Spinner inside the card."
-			},
-			{
-				"id": "ready",
-				"when": "All queries resolved",
-				"renders": "Greeting + quick actions + 4-stat row + 2 list cards."
-			},
-			{
-				"id": "empty-drafts",
-				"when": "drafts.records.length === 0",
-				"renders": "`No drafts yet. Start writing!` copy."
-			},
-			{
-				"id": "empty-comments",
-				"when": "comments.records.length === 0",
-				"renders": "`Inbox zero. Nothing pending.` copy."
-			}
-		],
-		"interactions": [
-			{
-				"trigger": "Click Write a post",
-				"effect": "`navigate('#/posts/new')`."
-			},
-			{
-				"trigger": "Click Add a page",
-				"effect": "`navigate('#/pages/new')`."
-			},
-			{
-				"trigger": "Click Upload media",
-				"effect": "`navigate('#/media')`."
-			},
-			{
-				"trigger": "Click a draft title",
-				"effect": "`navigate('#/posts/{id}/edit')`."
-			},
-			{
-				"trigger": "Click Moderate all",
-				"effect": "`navigate('#/comments')`."
-			}
-		],
-		"a11y": {
-			"focus-management": "Native heading + button order."
-		},
-		"constraints": [
-			{
-				"concern": "useentityrecord-null-userid",
-				"note": "`useEntityRecord('root','user', userId || 0)` is null-safe: passing `0` short-circuits the request without unmounting the hook. Keeping the hook present preserves React's hook-order invariant."
-			},
-			{
-				"concern": "totalitems-from-perpage-1",
-				"note": "WordPress core has no REST equivalent for `wp_count_posts` / `wp_count_comments`. Per-status `?per_page=1` requests + reading `X-WP-Total` (exposed by core-data as `totalItems`) is the documented work-around."
-			},
-			{
-				"concern": "title-shape",
-				"note": "Draft titles read `draft.title?.raw || draft.title?.rendered` with a `(no title)` fallback — REST returns title as `{ raw, rendered }` in edit context, and the latter is HTML-decoded so raw is preferred."
-			},
-			{
-				"concern": "no-welcome-panel",
-				"note": "WordPress's `Welcome` panel + `Site Health` widget + `Events and News` widget are NOT ported. The shell dashboard is intentionally smaller — see `docs/screens/dashboard-home.md` for the wp-admin equivalent."
-			},
-			{
-				"concern": "grid-fallback",
-				"note": "Uses `__experimentalGrid` — no WPDS 0.12 port."
-			}
-		],
-		"design-system-leakage": [
-			{
-				"import": "@wordpress/ui#Button, Card, Stack, Text",
-				"purpose": "Card layout, stat values, action buttons."
-			},
-			{
-				"import": "@wordpress/components#__experimentalGrid, Spinner",
-				"purpose": "4-column stat grid + 2-column card row + loading spinners — WPDS 0.12 gaps."
-			}
-		]
-	}
-}
